requirements.txt
pandas>=1.5.0
numpy>=1.24.0
yfinance>=0.2.18
scikit-learn>=1.3.0
tensorflow>=2.13.0
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
redis>=5.0.0
plotly>=5.17.0
ta>=0.10.2
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
